% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cap_fig}
\alias{cap_fig}
\title{Figure caption generator}
\usage{
cap_fig(name, cite = FALSE, caption = "")

cap_fig(name, cite = FALSE, caption = "")
}
\description{
Allows you to create a table caption with automatically generated numbered prefix.

Allows you to create a table caption with automatically generated numbered prefix.
}
\examples{
my_caption <- cap_fig("flower_plot", "Histogram of flower colors.")
my_caption <- cap_tab("flower_table", "Frequency table of flower colors.")
}
\seealso{
\code{\link{captioner}}, \code{\link{cap_tab}}

\code{\link{captioner}}, \code{\link{cap_fig}}
}

