% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_knitr-internal.R
\name{cap_knitr}
\alias{cap_knitr}
\title{cap_knitr}
\usage{
cap_knitr(where = "after")
}
\arguments{
\item{before}{Logical determining whether to display the caption before
or after the figure.}
}
\description{
Option to automatically show a caption after a knitr chunk.
}
\details{
cap_knitr sets the knitr option \code{cap} so your captions will
automatically display.  The caption is found by accessing the chunk name,
so the caption and chunk names must match. \code{cap} takes the name of
the corresponding captioner function, in quotes.
}
\keyword{internal}

