% Generated by roxygen2 (4.0.2): do not edit by hand
\name{caption}
\alias{caption}
\title{Caption generator}
\usage{
caption(...)
}
\arguments{
\item{...}{Character string(s) containing the caption}
}
\value{
A character string containing the caption
}
\description{
Works well in conjunction with \code{\link{incrementer}} for creating a numbered figure
caption.  The caption can be stored as an object and used in .Rmd documents for numbering
and labelling figures.
}
\examples{
caption("This histogram shows the distribution of flower species.")

fig_nums <- incrementer()
caption(fig_nums("tree_histogram", label = TRUE),
        "This histogram shows the distribution of tree species.")
}
\seealso{
\code{\link{incrementer}}
}

