% Generated by roxygen2 (4.0.2): do not edit by hand
\name{captioner}
\alias{captioner}
\title{Captioner function}
\usage{
captioner(prefix = "Figure")
}
\arguments{
\item{prefix}{Character string containing text to go before object number}

\item{name}{Character string containing a unique object name}

\item{cite}{Logical indicating whether or not you would like to generate a short form to use for in-text citation}

\item{caption}{The object caption}
}
\value{
A captioner function.

Captioner generates a function with the following parameters

A character string containing the prefix and object number with caption

A character string containing the prefix and object number with caption if \code{cite = TRUE}
}
\description{
Creates a function to create numbered captions for figures, tables or other objects
}
\examples{
fig_nums <- captioner()
fig_nums("flower_plot", "This plot shows the distribution of flower colors")
fig_nums("flower_plot", cite = TRUE)
}

