% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_class-internal.R
\name{check_class}
\alias{check_class}
\title{Check the class of an input parameter}
\usage{
check_class(.param, .class)
}
\arguments{
\item{.param}{The object that you want to check the class of}

\item{.class}{The expected class for \code{.param}}
}
\description{
Check the class of an input parameter
}
\keyword{internal}

