% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/increment-internal.R
\name{increment}
\alias{increment}
\title{Increment function, for incrementing a hierarchical number}
\usage{
increment(values, level)
}
\description{
Increment function, for incrementing a hierarchical number
}

