% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cap_fig}
\alias{A}
\alias{cap_fig}
\alias{cap_tab}
\alias{captioner}
\alias{captions}
\alias{create}
\alias{figure}
\alias{generating}
\alias{numbered}
\alias{or}
\alias{own}
\alias{simplified}
\alias{table}
\alias{to}
\alias{without}
\alias{your}
\title{Simple figure/table caption generators}
\usage{
cap_fig(name, cite = FALSE, caption = "")
cap_tab(name, cite = FALSE, caption = "")
}
\arguments{
\item{name}{Character string giving a unique name for the figure or table}

\item{cite}{Logical indicating whether or not you would like a short form for in text citing}

\item{caption}{Character string containing the caption}
}
\value{
Character string containg the prefix and figure/table number with or without caption
}
\description{
Simple figure/table caption generators
}
\examples{
cap_fig("flower_plot", "Histogram of flower colors.")
cap_tab("flower_table", "Table showing flower color distribution.")
}
\seealso{
\code{\link{captioner}}
}

