% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{captioner}
\alias{captioner}
\alias{captioner-package}
\title{captioner: A package for numbering figures and generating captions}
\description{
captioner: A package for numbering figures and generating captions
}
\author{
Alathea D Letaw, \email{alathea@zoology.ubc.ca}
}

