% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{captioner}
\alias{captioner}
\alias{captioner-package}
\title{captioner: A package for numbering figures and generating captions}
\description{
Contains the functions \code{\link{caption}} for generating captions
and \code{\link{incrementer}} for generating figure or table
numbers.
}
\author{
Alathea D Letaw, \email{alathea@zoology.ubc.ca}
}

