% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captioner.R
\name{captioner}
\alias{captioner}
\title{Captioner function}
\usage{
captioner(prefix = "Figure", suffix = ": ", style = NULL,
  style_prefix = FALSE, levels = 1, type = NULL, infix = ".",
  before = FALSE, knitr_op = NULL, css_class = NULL)
}
\arguments{
\item{prefix}{Character string containing text to go before object number.
The default is \emph{Figure}.}

\item{suffix}{Character string containing text to go after object number and
before caption. The default is ": ".}

\item{style}{Optional character string indicating md style to use. Possible
options: bold \emph{b}, italic \emph{i}, or bold+italic \emph{bi}.}

\item{style_prefix}{Logical specifying whether the style should be applied to
the prefix only (\emph{TRUE}) or the entire caption.  The default is \emph{FALSE}.}

\item{levels}{Logical or number indicating whether or not you want hierarchical
numbering, and if so, how many levels.  Hierarchical numbering is turned off by
default.}

\item{type}{Vector with same length as \code{levels} indicating whether figure
numbering should be numeric (\emph{n}), lowercase character (\emph{c}), or
uppercase character (\emph{C}). If unspecified, \code{captioner} will revert to
all numeric values.}

\item{infix}{Character string containing text to go between figure numbers if
hierarchical numbering is on.  Default is \emph{.}}

\item{before}{Logical indicating whether to display the caption before or after the
figure.  Applies only to automatic caption display (e.g. with a hook).}

\item{knitr_op}{A named list containing any other chunk options desired.}

\item{css_class}{Assign a css class to the caption. Places the caption into 
a span html element with a class.}
}
\value{
A captioner function.
}
\description{
Creates a function to create numbered captions for figures, tables or other objects
}
\details{
Captioner generates a function with the following parameters

\code{name}: Character string containing a unique object name

\code{caption}: The object caption

\code{display}: Character string (or logical) indicating what display mode you would like:
\itemize{
\item \emph{full} or \emph{f} is the default and shows the entire caption with prefix and number
\item \emph{cite} or \emph{c} displays just the prefix and number, without the caption
\item \emph{num} or \emph{n} displays just the number
}

\code{level}: Optional numeric used to bump up the numbering if you have hierarchical numbering.  See also \code{\link{bump}}.
  
\code{cite}: Deprecated.  Please use \code{display}.

\code{num}: Deprecated. Please use \code{display}.

And returns a character string containing the prefix and object number with or without a caption.
The initial numbering is determined based on the order of caption creation.  However, this order
is modified based on the citations you use.  The first figure to be cited will be moved to the
beginning of the list, becoming \emph{Figure 1}.
}
\examples{
fig_nums <- captioner()
fig_nums("flower_plot", "This plot shows the distribution of flower colors")
fig_nums("flower_plot", display = "cite")
fig_nums("flower_plot")

tables <- captioner(prefix = "Table", levels = 2)
tables("a", "Table of world populations sorted from greatest to least.")
  
}

