% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/captioner.R
\name{captioner}
\alias{captioner}
\title{Captioner function}
\usage{
captioner(prefix = "Figure", auto_space = TRUE, levels = 1, type = NULL,
  infix = ".", link = FALSE,
  fmt = as.character(knitr::opts_knit$get()[["rmarkdown.pandoc.to"]]))
}
\arguments{
\item{prefix}{Character string containing text to go before object number. The default is "Figure".}

\item{auto_space}{Logical indicating whether or not a space should automatically be added following the prefix.  Space is added by default.}

\item{levels}{Logical or number indicating whether or not you want hierarchical numbering, and if so, how many levels.  Hierarchical numbering is turned off by default.}

\item{type}{Vector with same length as `levels` indicating whether figure numbering should be numeric ("n"), lowercase character ("c"), or uppercase character ("C").  If unspecified, `captioner` will revert to all numeric values.}

\item{infix}{Character string containing text to go between figure numbers if hierarchical numbering is on.  Default is "."}

\item{link}{Logical indicating whether you want citations linked to figures for HTML or PDF output. Default is FALSE.}
}
\value{
A captioner function.
}
\description{
Creates a function to create numbered captions for figures, tables or other objects
}
\details{
Captioner generates a function with the following parameters

\code{name}: Character string containing a unique object name

\code{caption}: The object caption

\code{display}: Character string (or logical) indicating what display mode you would like: "full" or "f" is the default and shows the entire caption with prefix and number; "cite" or "c" displays just the prefix and number, without the caption; "num" or "n" displays just the number

\code{level}: Optional numeric used to "bump" up the numbering if you have hierarchical numbering.  See also \code{\link{bump}}.

\code{cite}: Deprecated.  Please use \code{display}.

\code{num}: Deprecated. Please use \code{display}.

And returns a character string containing the prefix and object number with or without a caption.
The initial numbering is determined based on the order of caption creation.  However, this order
is modified based on the citations you use.  The first figure to be cited will be moved to the
beginning of the list, becoming "Figure 1".

If \code{link = TRUE}, then citations are hyperlinked to the referenced
caption. This only applies when rendering an Rmarkdown document to HTML or
PDF (see \url{http://rmarkdown.rstudio.com}). For the link to work, the
caption needs to be output using the option \code{fig_caption: yes} in the
front matter of the Rmarkdown document. Since for PDF output, figure
numbering is internally handled, this number gets used over the captioner
generated number so special use of \code{level} does not get applied.
}
\examples{
\donttest{
# Create a new captioner object:
fig_nums <- captioner()

# Store a caption with the key word "flower_plot". Print the caption
# including the figure number:
fig_nums("flower_plot", "This plot shows the distribution of flower colors")

# Cite the figure e.g. "Figure 1", using the key word:
fig_nums("flower_plot", cite = TRUE)

# Now you can print the caption any time using the designated key:
fig_nums("flower_plot")

# Create a captioner object with hierarchical numbering:
tables <- captioner(prefix = "Table", levels = 2)
tables("a", "Table of world populations sorted from greatest to least.")
}

# Create a captioner object with links (for HTML or PDF output):
fig_nums <- captioner(link = TRUE)
}

