% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/captioner.R
\name{captioner}
\alias{captioner}
\title{Captioner function}
\usage{
captioner(prefix = "Figure", auto_space = TRUE, levels = 1, type = NULL,
  infix = ".")
}
\arguments{
\item{prefix}{Character string containing text to go before object number. The default is "Figure".}

\item{auto_space}{Logical indicating whether or not a space should automatically be added following the prefix.  Space is added by default.}

\item{levels}{Logical or number indicating whether or not you want hierarchical numbering, and if so, how many levels.  Hierarchical numbering is turned off by default.}

\item{type}{Vector with same length as `levels` indicating whether figure numbering should be numeric ("n"), lowercase character ("c"), or uppercase character ("C").}

\item{infix}{Character string containing text to go between figure numbers if hierarchical numbering is on.}
}
\value{
A captioner function.
}
\description{
Creates a function to create numbered captions for figures, tables or other objects
}
\details{
Captioner generates a function with the following parameters

\code{name}: Character string containing a unique object name

\code{caption}: The object caption

\code{cite}: Logical indicating whether or not you would like to generate a short form to use for in-text citation

And returns a character string containing the prefix and object number with or without a caption.
The initial numbering is determined based on the order of caption creation.  However, this order
is modified based on the citations you use.  The first figure to be cited will be moved to the
beginning of the list, becoming "Figure 1".
}
\examples{
Create a new captioner object:
fig_nums <- captioner()

Store a caption with the key word "flower_plot". Print the caption
including the figure number:

fig_nums("flower_plot", "This plot shows the distribution of flower colors")

Cite the figure e.g. "Figure 1", using the key word:
fig_nums("flower_plot", cite = TRUE)

Now you can print the caption any time using the designated key:
fig_nums("flower_plot")
}

