% Generated by roxygen2 (4.0.2): do not edit by hand
\name{captioner}
\alias{captioner}
\title{Captioner function}
\usage{
captioner(prefix = "Figure")
}
\arguments{
\item{prefix}{Character string containing text to go before object number}
}
\value{
A captioner function.
}
\description{
Creates a function to create numbered captions for figures, tables or other objects
}
\details{
Captioner generates a function with the following parameters

\code{name}: Character string containing a unique object name

\code{cite}: Logical indicating whether or not you would like to generate a short form to use for in-text citation

\code{caption}: The object caption

And returns either character string containing the prefix and object number with or without a caption
}
\examples{
fig_nums <- captioner()
fig_nums("flower_plot", "This plot shows the distribution of flower colors")
fig_nums("flower_plot", cite = TRUE)
}

