% Generated by roxygen2 (4.0.2): do not edit by hand
\name{increment}
\alias{increment}
\title{Increment function, for incrementing a hierarchical number}
\usage{
increment(values, level)
}
\arguments{
\item{values}{The list of figure numbers that need to be incremented}

\item{level}{Which number should be incremented}
}
\description{
Increment function, for incrementing a hierarchical number
}

