% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/increment-internal.R
\name{increment}
\alias{increment}
\title{Increment function, for incrementing a hierarchical number}
\usage{
increment(values, level)
}
\arguments{
\item{values}{The list of figure numbers that need to be incremented}

\item{level}{Which number should be incremented}
}
\description{
Increment function, for incrementing a hierarchical number
}
\keyword{internal}

