% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incrementer}
\alias{incrementer}
\title{Object incrementer function}
\usage{
incrementer()
}
\arguments{
\item{obj}{Character string containing a unique object name}

\item{label}{Logical indicating whether or not you would like to generate a label}

\item{pre_text}{Character string containing text to go before your object number}
}
\value{
An incrementer function.

A character string containing the figure label and number, if \code{label = TRUE}
}
\description{
Creates a function to keep track of and number figures, tables or other objects.
Incrementer takes no arguments but generates a function with the parameters found below.
}
\examples{
fig_nums <- incrementer()
fig_nums("flower_plot")
fig_nums("tree_plot", label = TRUE)
}
\seealso{
\code{\link{caption}}
}

